rune/libenclave/attestation/sgx: rename the sgx_challenger directory as challenger
